# lib/api-zod/src/generated/types/tradesResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Lukas AI Agent + Higgsfield API
 * OpenAPI spec version: 0.1.0
 */
import type { TradeRecord } from './tradeRecord';

export interface TradesResponse {
  trades: TradeRecord[];
  total: number;
}