# lib/api-zod/src/generated/types/tradesSummary.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Lukas AI Agent + Higgsfield API
 * OpenAPI spec version: 0.1.0
 */
import type { TradeRecord } from './tradeRecord';

export interface TradesSummary {
  open_count: number;
  closed_count: number;
  total_count: number;
  realized_pnl: number;
  win_count: number;
  loss_count: number;
  win_rate: number;
  open_positions: TradeRecord[];
}